chore: add debug code to see why it breaks
test(..) function works fine but the same code breaks in the caller.
It appears the function is too large and we fail to inline quat_from_matrix.
quat_from_matrix is inline and doesn't use __vectorcall, it calls an implementation that uses __vectorcall.
The inner function returns the values in xmm0 and xmm1 as expected but quat_from_matrix fails
to read them and write them on the stack where the caller expects them.
When quat_from_matrix is inlined, there is no issue.
In test(..), it is inlined and it works. In the caller, it isn't and it fails the unit test.