refactor(scalar): make SSE2 scalarf and scalard concrete types
With SSE2, we use a concrete type for scalarf/scalard unlike other platforms and other types
like vector4f and quatf. We don't use a concrete type when we can avoid it to help the compiler
optimize as much as possible. But we must be able to tell a scalar apart from a vector for
return type overloading and argument overloading.

For example, we want to support vector_mul(vec4, vec4) and vector_mul(vec4, scalar).
When scalarf is a 'float', the type is distinct and everything works as expected
but if we use __m128, the type is the same as vector4f and we won't be able to tell
them apart.

Another example is vector_dot where we want to support returning a float, a scalarf, and
a vector4f depending on what the user expects. We could always return a float/scalarf but
if we need a vector4f it is less efficient if _mm_dp_ps is used: we would have an extra
shuffle.

Using a concrete type here allows us to tell the types apart and properly overload them
when required. The compiler should still be able to optimize properly.