fix: disable MSVC security buffer checks
Because scalarf and other types are wrapped by structs, sometimes MSVC generates
security buffer checks when /GS is used. That option is also enabled by default and
required on the XB1. This can cause functions as short as 8 instructions to fail to inline
because the security check adds another 8 instruction. This can significantly degrade
performance in math heavy and hot code. Not every function is vulnerable to this but
as a safety precaution, it will be set on every function. RTM code is very simple to
manually validate for stack safety, we do not need the compiler's help.